split filecontract stuff into its own file in types package
